requirements.txt
# 核心ML依赖
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.3

# MLOps工具
mlflow==2.8.1

# 开发工具
python-dotenv==1.0.0
pytest==7.4.0
black==23.7.0
flake8==6.0.0
isort==5.12.0